+++ packages/catalog-model/examples/apis/backstage-mcp-server-api.yaml
@@ -1,4 +1,4 @@
-apiVersion: backstage.io/v1alpha2
+apiVersion: backstage.io/v1alpha1
 kind: API
 metadata:
   name: backstage-mcp-server
