+++ packages/catalog-model/src/kinds/index.ts
@@ -20,16 +20,10 @@ export type {
   ApiEntityV1alpha1,
 } from './ApiEntityV1alpha1';
 export {
-  apiEntityV1alpha2Validator,
   isMcpServerApiEntity,
-  mcpServerApiEntityV1alpha2Validator,
-} from './ApiEntityV1alpha2';
-export type {
-  ApiEntityV1alpha2,
-  ApiEntityV1alpha2Default,
-  McpServerApiEntityV1alpha2,
-  McpServerRemote,
-} from './ApiEntityV1alpha2';
+  mcpServerApiEntityValidator,
+} from './McpServerApiEntity';
+export type { McpServerApiEntity, McpServerRemote } from './McpServerApiEntity';
 export { componentEntityV1alpha1Validator } from './ComponentEntityV1alpha1';
 export type {
   ComponentEntityV1alpha1 as ComponentEntity,
