+++ packages/catalog-model/src/model/defaultCatalogEntityModel.ts
@@ -15,7 +15,7 @@
  */
 
 import { apiEntityModel } from '../kinds/ApiEntityV1alpha1';
-import { apiEntityV1alpha2Model } from '../kinds/ApiEntityV1alpha2';
+import { mcpServerApiEntityModel } from '../kinds/McpServerApiEntity';
 import { componentEntityModel } from '../kinds/ComponentEntityV1alpha1';
 import { domainEntityModel } from '../kinds/DomainEntityV1alpha1';
 import { groupEntityModel } from '../kinds/GroupEntityV1alpha1';
@@ -37,7 +37,7 @@ export const defaultCatalogEntityModel = createCatalogModelLayer({
   layerId: 'catalog.backstage.io/default-entity-model',
   builder: model => {
     model.import(apiEntityModel);
-    model.import(apiEntityV1alpha2Model);
+    model.import(mcpServerApiEntityModel);
     model.import(componentEntityModel);
     model.import(domainEntityModel);
     model.import(groupEntityModel);
