+++ packages/catalog-model/src/schema/kinds/API.v1alpha1.mcp-server.schema.json
@@ -1,10 +1,10 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema",
-  "$id": "ApiV1alpha2McpServer",
+  "$id": "ApiV1alpha1McpServer",
   "description": "An MCP (Model Context Protocol) server exposed as an API entity. See RFC backstage/backstage#32062.",
   "examples": [
     {
-      "apiVersion": "backstage.io/v1alpha2",
+      "apiVersion": "backstage.io/v1alpha1",
       "kind": "API",
       "metadata": {
         "name": "backstage-mcp-actions",
@@ -33,7 +33,7 @@
       "required": ["spec"],
       "properties": {
         "apiVersion": {
-          "enum": ["backstage.io/v1alpha2"]
+          "enum": ["backstage.io/v1alpha1", "backstage.io/v1beta1"]
         },
         "kind": {
           "enum": ["API"]
